pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study-agent"
version = "0.1.0"
description = "Local PC study monitoring agent MVP"
requires-python = ">=3.11"
dependencies = [
  "mss==10.1.0",
  "opencv-python==4.11.0.86",
  "Pillow==11.1.0",
  "requests==2.32.3",
  "python-dotenv==1.0.1",
  "pydantic==2.11.1",
]

[project.scripts]
study-agent = "study_agent.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
